feat: better payments api

+++ src/main/java/com/manager/payments/adapter/out/persistence/payments/PaymentRepositoryAdapter.java
@@ -1,8 +1,11 @@
 package com.manager.payments.adapter.out.persistence.payments;
 
 import com.manager.payments.application.port.out.PaymentRepository;
+import com.manager.payments.model.exceptions.PaymentNotFoundException;
 import com.manager.payments.model.payments.Payment;
 import com.manager.payments.model.payments.PaymentStatus;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Repository;
 
 import java.time.LocalDate;
@@ -20,6 +23,12 @@ public class PaymentRepositoryAdapter implements PaymentRepository {
                                     PaymentJpaRepository paymentJpaRepository) {
         this.paymentMapper = paymentMapper;
         this.paymentJpaRepository = paymentJpaRepository;
+    }
+
+    @Override
+    public Page<Payment> findAll(Pageable pageable) {
+        Page<PaymentJpaEntity> paymentJpaEntities = paymentJpaRepository.findAll(pageable);
+        return paymentJpaEntities.map(paymentMapper::toPayment);
     }
 
     @Override
@@ -55,6 +64,9 @@ public class PaymentRepositoryAdapter implements PaymentRepository {
 
     @Override
     public void deleteById(UUID id) {
+        if (!paymentJpaRepository.existsById(id))
+            throw new PaymentNotFoundException(id);
+
         paymentJpaRepository.deleteById(id);
     }
